Update pal from commit: 59416f33
* Add ReadSetting cross-platform wrapper
* Correct RereadSetting name
* dEQP-VK.dynamic_rendering.*cmdbuffer* - failing on Linux
* Incorrect CPU logical cores shown in the RGP Device Configuration
* WriteBufferImmediate  CP copydata top: ME->PFP and Cleanup
* Implementation of device info trace source
* Add option to open shared memory with globalGpuVa
* [Streaming SDK 1.5] Pass frame index from the app to KMD
* Debug logging support in PAL
* Fixup Linux ArchiveFile implementation
* Adjust default tessFactorBufferSizePerSe
* Tweak Shader Cache Hash Map settings to reduce memory usage
* StringView class implementation
*  Hook up memory budget value in PAL to SW addrlib
* Add ICmdBuffer::CmdCopyMemoryByGpuVa
* Fix settings script
* PAL_NOT_IMPLEMENTED Assert triggers with MGPU
* Fix for SPM regr
* Remove pointless cache layer alerts
* Null device dummy values for swizzling
* Add utility function to read a file into memory
* Use old SPM layout until RGP is ready
* Ubuntu21.04+Wayland Vulkan perf ~40% lower compared U20.04
* Add sample info to ShaderEarlyCompileInfo
* Fix InitBusAddressableGpuMemory decorator
* Updated waForce1kHtileMin size in Gfx9
* RDF lib update
* Clean up NullDevice tables

+++ src/gpuUtil/traceSession.cpp
@@ -188,7 +188,7 @@ Result TraceSession::CollectTrace(
         result = RdfResultToPalResult(rdfStreamSeek(m_pCurrentStream, 0));
     }
 
-    size_t streamSize;
+    int64 streamSize;
     if (result == Result::Success)
     {
         result = RdfResultToPalResult(rdfStreamGetSize(m_pCurrentStream, &streamSize));
@@ -213,7 +213,7 @@ Result TraceSession::CollectTrace(
             if (pData != nullptr)
             {
                 // Read all trace data in the current stream in RDF format
-                size_t bytesRead = 0;
+                int64 bytesRead = 0;
                 if (result == Result::Success)
                 {
                     result = RdfResultToPalResult(rdfStreamRead(m_pCurrentStream, streamSize, pData, &bytesRead));
